requirements.txt
gunicorn
asgiref
beautifulsoup4
bootstrap4
certifi
cffi
charset-normalizer
cryptography
defusedxml
Django
django-bootstrap4
django-filter
django-phonenumber-field
djangorestframework
idna
oauthlib
phonenumbers
Pillow
pycparser
PyJWT
python-social-auth
python3-openid
pytz
requests
requests-oauthlib
social-auth-app-django
social-auth-core
soupsieve
sqlparse
urllib3
whitenoise
